+++ src/cow-react/modules/wallet/web3-react/hooks/useWalletInfo.ts
@@ -2,7 +2,7 @@ import { useWeb3React } from '@web3-react/core'
 import { Web3Provider } from '@ethersproject/providers'
 import useENSName from '@src/hooks/useENSName'
 import { UNSUPPORTED_WC_WALLETS } from 'constants/index'
-import { getConnection } from '@cow/modules/wallet/web3-react/utils/connection'
+import { getConnection } from '@cow/modules/wallet/web3-react/types'
 import { SafeInfoResponse } from '@gnosis.pm/safe-service-client'
 import { ConnectionType } from '@cow/modules/wallet'
 import { useIsSmartContractWallet } from '@cow/common/hooks/useIsSmartContractWallet'
